requirements.txt
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.18.1
Werkzeug==3.0.3
